Fix issues mentioned in review
